+++ packages/design/jobberStyle.js
@@ -2,6 +2,12 @@
 const fs = require("fs");
 // eslint-disable-next-line import/no-internal-modules
 const customPropertiesObject = require("./src/foundation.js");
+// eslint-disable-next-line import/no-internal-modules
+const { getShadowStyles } = require("./src/getMobileShadows.js");
+// eslint-disable-next-line import/no-internal-modules
+const { getMobileLineHeights } = require("./src/getMobileLineHeights");
+// eslint-disable-next-line import/no-internal-modules
+const { getMobileFontSizes } = require("./src/getMobileFontSizes");
 
 const regexExpressions = {
   cssVars: /var\((.*)\)/,
@@ -25,9 +31,7 @@ fs.writeFile("./foundation.js", jsonContent, "utf8", function (err) {
     console.log("An error occured while writing JSON object to File.");
     return console.log(err);
   }
-  console.log("JSON file has been saved.");
 });
-
 const scssColors = getResolvedSCSSVariables(resolvedCssVars);
 
 fs.writeFile(
@@ -39,9 +43,9 @@ fs.writeFile(
       console.log("An error occured while writing SCSS to File.");
       return console.log(err);
     }
-    console.log("SCSS file has been saved.");
   },
 );
+writeMobileFoundationFiles();
 
 /**
  * Recursively resolve css custom properties.
@@ -258,4 +262,69 @@ function getPropertyValue(cssVar) {
 function removeNewLines(text) {
   if (!text) return text;
   return text.replace(/(\r\n|\n|\r)/gm, "");
+}
+
+function writeMobileFoundationFiles() {
+  const { androidShadows, iOSShadows } = getShadowStyles(resolvedCssVars);
+  const mobileLineHeightValues = getMobileLineHeights();
+  const mobileFontSizeValues = getMobileFontSizes();
+  const androidFoundationJobberStyle = {
+    ...resolvedCssVars,
+    ...androidShadows,
+    ...mobileLineHeightValues,
+    ...mobileFontSizeValues,
+  };
+  const iOSFoundationJobberStyle = {
+    ...resolvedCssVars,
+    ...iOSShadows,
+    ...mobileLineHeightValues,
+    ...mobileFontSizeValues,
+  };
+  const androidFoundationsExportString = `export const JobberStyle = ${JSON.stringify(
+    androidFoundationJobberStyle,
+    undefined,
+    2,
+  )}`;
+
+  const iOSFoundationsExportString = `export const JobberStyle = ${JSON.stringify(
+    iOSFoundationJobberStyle,
+    undefined,
+    2,
+  )}`;
+  fs.writeFile(
+    "./foundation.android.js",
+    androidFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing Android Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote Android Foundations file");
+    },
+  );
+  fs.writeFile(
+    "./foundation.ios.js",
+    iOSFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing iOS Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote iOS Foundations file");
+    },
+  );
+  fs.writeFile(
+    "./foundation.native.js",
+    iOSFoundationsExportString,
+    "utf-8",
+    err => {
+      if (err) {
+        console.log("An error occured while writing iOS Foundation File.");
+        return console.log(err);
+      }
+      console.log("Wrote Native Foundations file");
+    },
+  );
 }