+++ packages/components/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.10.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.9.4...@jobber/components@4.10.0) (2023-04-28)
+
+
+### Features
+
+* Icon map generator in 🔱 Atlantis! ([#1190](https://github.com/GetJobber/atlantis/issues/1190)) ([c19e800](https://github.com/GetJobber/atlantis/commit/c19e8007932309b16d00d4c3228be14450f7a3e1))
+
+
+
+
+
+## [4.9.4](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.9.3...@jobber/components@4.9.4) (2023-04-27)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.9.3](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.9.2...@jobber/components@4.9.3) (2023-04-27)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+## [4.9.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.9.1...@jobber/components@4.9.2) (2023-04-27)
+
+
+### Bug Fixes
+
+* **components:** calculate Z index for autocomplete  ([#1215](https://github.com/GetJobber/atlantis/issues/1215)) ([683e575](https://github.com/GetJobber/atlantis/commit/683e575bcbea6f2ac3f6e55a98cbb57065d69be7))
+
+
+
+
+
 ## [4.9.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.9.0...@jobber/components@4.9.1) (2023-04-26)
 
 **Note:** Version bump only for package @jobber/components